payment integration
Former-commit-id: 634fbd3ebc0f9a2c1ade8802db0d88b57e8b9250

+++ mobile/lib/Screens/Payment/components/body_invoice.dart
@@ -103,8 +103,8 @@ class Product {
 }
 
 class BodyInvoice extends StatelessWidget {
-  final double nominal;
-  final double points;
+  final int nominal;
+  final int points;
   final String namaMerchant;
   final String time;
   const BodyInvoice({
